2008-07-15 14:31 UTC+0200 Przemyslaw Czerpak (druzus/at/priv.onet.pl)   * harbour/contrib/hbmysql/Makefile   * harbour/contrib/hbodbc/Makefile   * harbour/contrib/hbsqlit3/Makefile   * harbour/contrib/hbfbird/Makefile   * harbour/contrib/hbcurl/Makefile   * harbour/contrib/hbhpdf/Makefile   * harbour/contrib/hbpgsql/Makefile   * harbour/contrib/hbfimage/Makefile   * harbour/contrib/hbgd/Makefile   * harbour/contrib/hbgf/hbgfgtk/Makefile     * do not add default system include directory to C_USR during testing       for foreign header files when HB_XBUILD variable is set - protection       against including wrong header files in cross builds
  * harbour/contrib/hbtpathy/ChangeLog
    * changed SVN EOL style to 'native'

  * harbour/contrib/hbbmcdx/bmdbfcdx.c
    * pacified warning

+++ harbour/contrib/hbbmcdx/bmdbfcdx.c
@@ -9071,7 +9071,7 @@ static ERRCODE hb_cdxCountScope( CDXAREAP pArea, void * pPtr, LONG * plRec )
    {
       LPBM_FILTER pMap = (LPBM_FILTER) pArea->dbfi.lpvCargo;
       if( pArea->dbfi.fFilter && pMap &&
-          !BM_GetBit( pMap->rmap, pMap->Size, *plRec ) )
+          !BM_GetBit( pMap->rmap, pMap->Size, ( ULONG ) *plRec ) )
       {
          *plRec = 0;
       }
